ZeinaGuard/sensor/requirements.txt
scapy
python-socketio
websocket-client
requests
rich
readchar
